relocate query param funcs to params pkg
