+++ pkg/proxy/params/params_test.go
@@ -17,10 +17,15 @@
 package params
 
 import (
+	"bytes"
 	"fmt"
+	"io/ioutil"
+	"net/http"
 	"net/url"
 	"reflect"
 	"testing"
+
+	"github.com/tricksterproxy/trickster/pkg/proxy/headers"
 )
 
 func TestUpdateParams(t *testing.T) {
@@ -42,6 +47,64 @@ func TestUpdateParams(t *testing.T) {
 	UpdateParams(params, updates)
 	if !reflect.DeepEqual(params, expected) {
 		fmt.Printf("mismatch\nexpected: %v\n     got: %v\n", expected, params)
+	}
+
+}
+
+func TestGetSetRequestValues(t *testing.T) {
+
+	const params = "param1=value1"
+
+	r, _ := http.NewRequest(http.MethodGet, "http://example.com/?"+params, nil)
+
+	v, s, hb := GetRequestValues(r)
+	if len(v) != 1 {
+		t.Errorf("expected %d got %d", 1, len(v))
+	}
+	if s != params {
+		t.Errorf("expected %s got %s", params, s)
+	}
+	if hb {
+		t.Errorf("expected false")
+	}
+
+	v.Set("param2", "value2")
+	SetRequestValues(r, v)
+	v, s, hb = GetRequestValues(r)
+	if len(v) != 2 {
+		t.Errorf("expected %d got %d", 2, len(v))
+	}
+	if s == params || s == "" {
+		t.Errorf("expected %s got %s", params+"&param2=value2", s)
+	}
+	if hb {
+		t.Errorf("expected false")
+	}
+
+	r, _ = http.NewRequest(http.MethodPost, "http://example.com/", ioutil.NopCloser(bytes.NewBufferString(params)))
+	r.Header.Set(headers.NameContentType, headers.ValueXFormURLEncoded)
+	v, s, hb = GetRequestValues(r)
+	if len(v) != 1 {
+		t.Errorf("expected %d got %d", 1, len(v))
+	}
+	if s != params {
+		t.Errorf("expected %s got %s", params, s)
+	}
+	if !hb {
+		t.Errorf("expected true")
+	}
+
+	v.Set("param2", "value2")
+	SetRequestValues(r, v)
+	v, s, hb = GetRequestValues(r)
+	if len(v) != 2 {
+		t.Errorf("expected %d got %d", 2, len(v))
+	}
+	if s == params || s == "" {
+		t.Errorf("expected %s got %s", params+"&param2=value2", s)
+	}
+	if !hb {
+		t.Errorf("expected true")
 	}
 
 }
